hash finalizado.

+++ f_hash.c
@@ -67,9 +67,9 @@ uint32_t lookup3 (
   return c;
 }
 
-unsigned int fhash(const char *word, unsigned int hashTableSize){
+size_t fhash(const char *word, size_t hashTableSize){
   unsigned int initval;
-  unsigned int hashAddress;
+  size_t hashAddress;
 
   initval = 12345;
   hashAddress = lookup3(word, strlen(word), initval);
